+++ Neuro/Initializers/GlorotUniform.cs
@@ -5,23 +5,23 @@ namespace Neuro.Initializers
 {
     public class GlorotUniform : InitializerBase
     {
-        public GlorotUniform(double gain = 1)
+        public GlorotUniform(float gain = 1)
         {
             Gain = gain;
         }
 
-        public static double NextDouble(int fanIn, int fanOut, double gain)
+        public static float NextSingle(int fanIn, int fanOut, float gain)
         {
-            double scale = 1 / Math.Max(1, (fanIn + fanOut) * 0.5);
-            double limit = Math.Sqrt(3 * scale);
-            return Uniform.NextDouble(-limit, limit);
+            float scale = 1 / (float)Math.Max(1, (fanIn + fanOut) * 0.5);
+            float limit = (float)Math.Sqrt(3 * scale);
+            return Uniform.NextSingle(-limit, limit);
         }
 
         public override void Init(Tensor t, int fanIn, int fanOut)
         {
-            t.Map(x => NextDouble(fanIn, fanOut, Gain), t);
+            t.Map(x => NextSingle(fanIn, fanOut, Gain), t);
         }
 
-        private readonly double Gain;
+        private readonly float Gain;
     }
 }
